break permissions function to components
